pyproject.toml
[project]
name = "template-langchain"
requires-python = ">=3.12.4"
version = "0.1.0"
description = ""
readme = "README.md"

dependencies = [
    # Core LangChain
    "langchain==0.3.7",
    "langchain-core==0.3.17",
    "langchain-community==0.3.7",
    "langchain-ollama==0.2.0",
    "langchain-postgres==0.0.16",
    "langgraph==0.2.34",
    
    # Database
    "psycopg2-binary==2.9.9",
    "pgvector==0.3.5", 
    "sqlalchemy==2.0.35",
    
    # Core utilities
    "python-dotenv==1.0.1",
    "pydantic==2.9.2",
    "requests==2.32.3",
    "typing-extensions==4.12.2",
    
    # Data processing
    "datasets==4.4.1",
    "pandas==2.2.3",

    # API
    "requests==2.32.3",
    "geopy==2.4.1",

    # Interface
    "streamlit==1.51.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.7.4",
]

test = [
    "pytest==8.3.3",
    "pytest-cov==5.0.0",
]

notebooks = [
    "jupyter==1.1.1",
    "ipykernel==6.29.5",
    "matplotlib==3.9.2",
    "seaborn==0.13.2",
    "rich==13.9.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sources"]

[tool.ruff]
line-length = 130
target-version = "py311"

exclude = [
    ".git",
    ".venv",
    ".vscode",
    "__pycache__",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings  
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "N",      # pep8-naming
]

[tool.ruff.lint.per-file-ignores]
"sources/rag/application/*.py" = ["ALL"]
